refactor: split program module into implementation and program
The standard for Modula-2 defines an implementation module to be a
program module with an implementation keyword preceding the module
header. However, to make it clearer, I have modified the grammar by
separating the individual modules into: definition, implementation and
program, just as every single parser or compiler treats it.

The tests have also been refactored accordingly