feat: add block production
Program modules, as well as procedures, have a block production rule. As
such, it was convenient to begin with this, but a very basic one. It has
only the optional body part followed by the end keyword. This
modification has led to some refactoring of the code, which makes it
more readable and less repetitive, not to mention that it follows the
grammar from the book a little closer.