refactor: change const_expression production chain
In previous commits, I declared const_expression as being composed of an
expression and provided a comprehensive definition of the entire
expression chain. This expression chain is used in assignments and other
statements within the body of a module or procedure.

However, it's important to note that const_declaration employs a
distinct set of expressions that are allowed within constant
declarations. For instance, function calls are not allowed within these
expressions. Consequently, it's essential to understand that a
designator cannot be assigned to a constant, which justifies the need
for separate declarations for constants.

To align with these changes, I have refactored the tests to accommodate
the aforementioned adjustments.