feat: implement enumeration and subrange simple types
I have successfully implemented the productions for enumeration and
subrange types, yet I've encountered a challenge while dealing with
subrange types. The issue arises when Modula-2 allows for real literals
like '45.' – this creates a conflict with the range ('..') operator. The
problem is that we must anticipate whether the '.' belongs to a range
token or is part of a real literal.

As of now, I haven't discovered a solution to this conflict within
TreeSitter. It seems that a custom scanner will be necessary to
distinguish between these two tokens accurately.

Please note that the test case for subrange types will remain unresolved
until I address this issue.